playlist browser: use terminalio
We discovered that RAM usage is ~100 bytes per character displayed in an
adafruit_display_text.label.Label.  There's not an easy fix for this,
and it was causing MemoryErrors in practice.  However, we can lower memory
usage by using terminalio.Terminal.  This requires us to switch to the
built-in font, but that's an acceptable compromise.